fix(web): share SavedCompare run mapping; fix serverMetrics drop in PDF export
The HydratedBenchmarkRef→ReportRun mapping was copy-pasted into both
ReportDetailPage and ReportPreviewPage. The earlier serverMetrics fix only
touched the detail page, so the print/export preview (the PDF surface) still
dropped serverMetrics — prefix-cache-hit / top-pod-share figures rendered
"data unavailable" in exported reports.

Extract a single `toReportRuns()` helper used by both surfaces (and dedupe the
twice-copied extractParamsSummary) so the two can no longer drift. The helper
includes serverMetrics ?? null.
